change final page imports to use new packages
