+++ packages/component-login/client/pages/LandingRedirectPage/index.js
