+++ packages/component-login/client/pages/LogoutPage/index.js
